introduce CurrentOrganizationType and rename SafeOrganizationType
