+++ spec/graphql/mutations/register_user_spec.rb
@@ -15,7 +15,6 @@ RSpec.describe Mutations::RegisterUser, type: :graphql do
           organization {
             id
             name
-            apiKey
           }
           membership {
             id
